+++ src/Polysemy/Internal/Kind.hs
@@ -1,6 +1,8 @@
+{-# OPTIONS_HADDOCK not-home #-}
+
 module Polysemy.Internal.Kind where
 
-import Data.Kind
+import Data.Kind (Type)
 
 ------------------------------------------------------------------------------
 -- | The kind of effects.
@@ -14,3 +16,7 @@ type Effect    = (Type -> Type) -> Type -> Type
 -- @since 0.5.0.0
 type EffectRow = [Effect]
 
+
+type family Append l r where
+  Append (a ': l) r = a ': (Append l r)
+  Append '[] r = r